add api service, use in container
